Update background.js
The main issue was in the showTimesheetNotification function. The notification options were missing, which was causing the error. I've made the following changes:

Added the required notification options:

type: 'basic'
iconUrl: 'icon.png'
title: 'Timesheet Reminder'
message: 'Don't forget to submit your timesheet!'
Added buttons for opening timesheets and snoozing
Set priority: 2 for high priority


Ensured that the iconUrl is set to 'icon.png', which should be available in your extension's directory.

These changes should resolve the error and allow the notification to be created successfully when the "Trigger Timesheet Reminder Now" button is clicked.